@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe ExternalIdentifier do
+describe ExternalIdentifier, type: :model do
 
   it { should belong_to :user }
 
